+++ vengine/src/vulkan_impl/components/device.h
@@ -27,13 +27,13 @@ class Device : public vstd::IOperatorNewBase {
 	void InitBindless();
 	mutable vstd::spin_mutex updateBindlessMtx;
 	mutable vstd::spin_mutex allocIdxMtx;
-	mutable vstd::vector<uint> bindlessIdx;
+	mutable vstd::vector<uint16> bindlessIdx;
 	mutable vstd::vector<VkWriteDescriptorSet> bindlessWriteRes;
 	mutable vstd::StackAllocator bindlessStackAlloc;
 
 public:
-	uint AllocateBindlessIdx() const;
-	void DeAllocateBindlessIdx(uint index) const;
+	uint16 AllocateBindlessIdx() const;
+	void DeAllocateBindlessIdx(uint16 index) const;
 	mutable vstd::DefaultMallocVisitor mallocVisitor;
 	vstd::unique_ptr<DescriptorSetManager> manager;
 	vstd::unique_ptr<GPUAllocator> gpuAllocator;
